+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2023-08-01
+
 ### Added
 
 - `ClientSet` now available from `prelude`.
@@ -121,7 +123,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release after separation from [lifescape](https://github.com/lifescapegame/lifescape).
 
-[unreleased]: https://github.com/lifescapegame/bevy_replicon/compare/v0.8.0...HEAD
+[unreleased]: https://github.com/lifescapegame/bevy_replicon/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/lifescapegame/bevy_replicon/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/lifescapegame/bevy_replicon/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/lifescapegame/bevy_replicon/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/lifescapegame/bevy_replicon/compare/v0.6.1...v0.7.0